tag flasher / timestamp content / bugfixes (#217)
- added webinterface for tag flasher
- added tcp transport for communicating with tag flasher (OTG USB also still works)
- added content 'timestamp', makes use of preloaded images and buttons on the 2.9" M3
- webinterface is now aware of C6 and flasher capabilities
- AP can run without ieee801.15.4 radio (i.e. flasher only) by shorting FLASHER_AP_TXD and FLASHER_AP_RXD
- added tcp transport option to OEPL-Flasher.py (serial also still works)
- added new environment OpenEPaperLink_Mini_AP_v4
- lots of finetuning and bug fixes

+++ ESP32_AP-Flasher/include/tag_db.h
@@ -55,7 +55,7 @@ class tagRecord {
 struct Config {
     uint8_t channel;
     char alias[32];
-    int16_t led;
+    uint8_t led;
     uint8_t tft;
     uint8_t language;
     uint8_t maxsleep;
